Renamed "map" package into "world"
