+++ src/main/java/com/omar/models/world/MapSize.java
@@ -1,4 +1,4 @@
-package com.omar.models.map;
+package com.omar.models.world;
 
 public enum MapSize {
     SMALL,
